Decode JWT in gateway
Rename 'local_server' to 'gateway' as this is how the team refers to
this part of the system.

Use JWTDecoder to extract the values from the JWT and verify the
signature.
